requirements.txt
Flask==2.0.2
gunicorn==20.1.0
werkzeug
marshmallow
google-cloud-storage
google-cloud-vision
google-cloud-texttospeech
google-cloud-translate
python-dotenv==0.15.0
python-docx
